Change mapFeature->plot in udfc query syntax
This was primarily done to make the django portions, client and server, more flexible. UDFCs are handled generically on the tiler, but the side sending the request is specific to plot, so it makes sense to send the query as "udf:plot" instead of "udf:mapFeature". If we start adding support for UDFCs on other mapFeatures, we need only add another rule to this config.