Only hide non-plot models when tree filter active
connects #75 on github

previously we were hiding non plot/tree models any time a tree was
included in the model list, or, in other words, whenever the specifity
of the search went past the underlying MapFeature.

The problem manifests on the client when 'unchecking' EmptyPlot for visibility we add a
show filter for 'Tree' and whatever other models should be shown. This
would trigger the condition to only show trees, making all other models
be hidden along with EmptyPlots.

The purpose for this original design decision was to account for the case when a
user enters some tree details to search by, but doesn't see the
checkboxes to hide other models and is confused by the results. They
probably just want to see tree results, so we do that for them. In order
to preserve this behavior, we gate specifically on whether a tree filter
was included, rather than looking at the display models.