Add support for restrictFeature clauses
It is possible to pass restrict clauses separate from displayFilters in
order to provide a limit based on eligible map feature types for
display, regardless of whether they were included in the show
filter. This is to enable a map to disable some of its map feature types
temporarily or permanently. Restrict clauses are a good fit for
configuring a static tiler URL for a map, while display filters are a
good fit for hashstring backed request parameters.