Add support searches on mft subclass models
connects #88 on github

In order to support searches on concrete model subclasses (Plot,
RainBarrel, etc.) of map feature, entries in the table config where
added that will (inner) join this table to the mapfeature table. This does two
things: adds the necessary columns for selection and filters out records
for other subclasses.

For the case of scalar udf searches on these models, the model-to-table
logic was modified to convert to mapfeature for all models besides tree,
(the one edge case), after validating that the model name is
valid. Since the inner join has already filtered the mapfeature table,
this is sufficient to return only relevant records.

No changes were required to support collection udfs.