Restore database URL parameter to support legacy tile clients
Older versions of the mobile apps still use tile URLs that contain
the database URL parameter, which is no longer used.

By restoring that URL paramter we restore support for those legacy clients

We however will continue to pull the database name from environment
settings, not from the URL parameter, which is now present but unused.