Restore missing "AS" portion of SQL select
connects #97 on github

This bug was introduced in 4097033. In this commit, DISTINCT statements
were removed for queries in which they were unnecessary. AS statements were
also removed in those cases. However AS statements are required for
Windshaft to correctly reference the geometry column, were also removed,
leading to invalid queries for polygons, or any field whose geometry
field is not the column name windshaft is looking for,
'the_geom_webmercator'.