Add instance configuration middleware to support unit conversion
Both mobile apps and the web app make tile requests. The most efficient way to
implement unit conversion support for all of these platforms is to add code to
the tiler.

All tile requests are associated with an instance, so we will always be able to
fetch the config object from the instance row.

I am implementing this lookup as middleware to ensure that the config is
retireved occurs before windshaft handles the request. I had initially tried
adding the lookup to `req2params` but this generated exceptions. Just inserting
a `setTimeout` into `req2params` produced the same exceptions, so it looks like
windshaft is expecting this method to be synchronous.

It is inefficient to look up the configuration for every tile request but:
- In my testing the query timings printed to the console in "debug mode" maxed
  out at 250ms and went as low as 5ms.
- This O(1) query will has the same cost regardless of the number of trees in an
  instance or, because of the index on `id`, the number of instances in the DB.
- Avoiding the need to invalidate a cache keeps the implementation as simple as
  possible.