Update moment to 2.19.3
Incorporates bug fixes.

The newer version of moment has a more relaxed default mode which was accepting
things like "1234 market st" as a valid date because it started with a 4 digit
number. Using strict mode resolves the issue, but requires the introduction of a
new format to correctly cover the two supported datetime representations.